+++ smac/facade/smac_ac_facade.py
@@ -29,6 +29,7 @@ from smac.initial_design.sobol_design import SobolDesign
 
 # intensification
 from smac.intensification.intensification import Intensifier
+from smac.intensification.successive_halving import SuccessiveHalving
 from smac.intensification.abstract_racer import AbstractRacer
 # optimizer
 from smac.optimizer.smbo import SMBO
@@ -488,10 +489,21 @@ class SMAC4AC(object):
         }
         if scenario.run_obj == 'quality':
             r2e_def_kwargs.update({
-                'success_states': [StatusType.SUCCESS, StatusType.CRASHED],
+                'success_states': [StatusType.SUCCESS, StatusType.CRASHED, StatusType.MEMOUT],
                 'impute_censored_data': False,
                 'impute_state': None,
             })
+
+        if isinstance(intensifier_instance, SuccessiveHalving) and scenario.run_obj == "quality":
+            r2e_def_kwargs.update({
+                'success_states': [StatusType.SUCCESS, StatusType.CRASHED,
+                                   StatusType.MEMOUT, StatusType.DONOTADVANCE,
+                                   ],
+                'consider_for_higher_budgets_state': [StatusType.DONOTADVANCE, StatusType.TIMEOUT,
+                                                      StatusType.CRASHED, StatusType.MEMOUT,
+                                                      ],
+            })
+
         if runhistory2epm_kwargs is not None:
             r2e_def_kwargs.update(runhistory2epm_kwargs)
         if runhistory2epm is None: